+++ handwrytten/__init__.py
@@ -18,7 +18,7 @@ Usage:
     ... )
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 from handwrytten.client import Handwrytten
 from handwrytten.exceptions import (